requirements.txt
langchain
langchain-openai
streamlit
langchain-community
PyPDFLoader
yt-dlp
pydub
atlassian-python-api
newspaper4k
youtube-transcript-api
chromadb
pandasai
tiktoken
crewai
newspaper4k